Cleanup of model timing/complexity/builder code
-remove (essentially useless) loggers from models
-move timing code from benchmark.py to net.py
-remove obsolete complexity code from metrics.py
-move remaining stuff from metrics.py to meters.py
-train_net/test_net: switch to using new complexity() code
-train_net/test_net: additional cleanup
